+++ CloudShell.ResourceDefinitions.ReferenceProviders/ConfigurationStore/ConfigurationStoreResourceTypeProvider.cs
@@ -38,7 +38,8 @@ public sealed class ConfigurationStoreResourceTypeProvider :
                 ValueShape: new(ResourceAttributeValueKind.String)),
             [Attributes.EntryCount] = new(
                 DefaultValue: 0,
-                ValueShape: new(ResourceAttributeValueKind.Integer))
+                ValueShape: new(ResourceAttributeValueKind.Integer),
+                ReadOnly: true)
         },
         Operations:
         [